requirements.txt
youtube_transcript_api==0.6.1
sumy==0.11.0
nltk==3.8.1
numpy==1.25.2
textbase-client==0.1.4